+++ jobs/genome_mapping/aln_4.indel_realign.sh
@@ -16,23 +16,28 @@ source $(pwd)/$SM/run_info
 set -o nounset
 set -o pipefail
 
+DONE=$SM/run_status/aln_4.indel_realign.done
+
 printf -- "---\n[$(date)] Start RealignerTargetCreator.\n"
 
-$JAVA -Xmx58G -Djava.io.tmpdir=tmp -jar $GATK \
-    -T RealignerTargetCreator -nt $NSLOTS \
-    -R $REF -known $MILLS -known $INDEL1KG \
-    -I $SM/alignment/$SM.markduped.bam \
-    -o $SM/alignment/realigner.intervals
+if [[ -f $DONE ]]; then
+    $JAVA -Xmx58G -Djava.io.tmpdir=tmp -jar $GATK \
+        -T RealignerTargetCreator -nt $NSLOTS \
+        -R $REF -known $MILLS -known $INDEL1KG \
+        -I $SM/alignment/$SM.markduped.bam \
+        -o $SM/alignment/realigner.intervals
 
-printf -- "---\n[$(date)] Finish RealignerTargetCreator.\n"
-printf -- "---\n[$(date)] Start IndelRealigner.\n---\n"
+    printf -- "---\n[$(date)] Finish RealignerTargetCreator.\n"
+    printf -- "---\n[$(date)] Start IndelRealigner.\n---\n"
 
-$JAVA -Xmx58G -Djava.io.tmpdir=tmp -jar $GATK \
-    -T IndelRealigner \
-    -R $REF -known $MILLS -known $INDEL1KG \
-    -targetIntervals $SM/alignment/realigner.intervals \
-    -I $SM/alignment/$SM.markduped.bam \
-    -o $SM/alignment/$SM.realigned.bam
-rm $SM/alignment/$SM.markduped.{bam,bai} $SM/alignment/realigner.intervals
+    $JAVA -Xmx58G -Djava.io.tmpdir=tmp -jar $GATK \
+        -T IndelRealigner \
+        -R $REF -known $MILLS -known $INDEL1KG \
+        -targetIntervals $SM/alignment/realigner.intervals \
+        -I $SM/alignment/$SM.markduped.bam \
+        -o $SM/alignment/$SM.realigned.bam
+    rm $SM/alignment/$SM.markduped.{bam,bai} $SM/alignment/realigner.intervals
+    touch $DONE
+fi
 
 printf -- "[$(date)] Finish IndelRealigner.\n---\n"
