+++ jobs/genome_mapping/pre_1.download.sh
@@ -18,28 +18,36 @@ source $(pwd)/$SM/run_info
 set -o nounset
 set -o pipefail
 
+mkdir -p $SM/run_status
+DONE=$SM/run_status/pre_1.download.$FNAME.done
+
 printf -- "---\n[$(date)] Start download: $FNAME\n"
 
-mkdir -p $SM/downloads
+if [[ -f $DONE ]]; then
+    echo "Skip this step."
+else
+    mkdir -p $SM/downloads
 
-rc=0
-n=0
-until [[ $n -eq 5 ]]; do
-    if [[ $LOC =~ ^syn[0-9]+ ]]; then
-        $SYNAPSE get $LOC --downloadLocation $SM/downloads/ && { rc=$?; break; } || rc=$?
-    elif [[ $LOC =~ ^s3:.+ ]]; then
-        $AWS s3 ls $LOC || {
-            printf "Set an NDA AWS token\n\n"
-            eval "$($PIPE_HOME/utils/nda_aws_token.sh -r ~/.nda_credential)"
-        }
-        $AWS s3 cp --no-progress $LOC $SM/downloads/ && { rc=$?; break; } || rc=$?
-    else
-        ls -lh $LOC && ln -sf $(readlink -f $LOC) $SM/downloads/ || rc=$?
-        break
-    fi
-    n=$((n+1))
-    printf "Download try $n failed.\n\n"
-done
-[[ $rc -eq 0 ]] || false
+    rc=0
+    n=0
+    until [[ $n -eq 5 ]]; do
+        if [[ $LOC =~ ^syn[0-9]+ ]]; then
+            $SYNAPSE get $LOC --downloadLocation $SM/downloads/ && { rc=$?; break; } || rc=$?
+        elif [[ $LOC =~ ^s3:.+ ]]; then
+            $AWS s3 ls $LOC || {
+                printf "Set an NDA AWS token\n\n"
+                eval "$($PIPE_HOME/utils/nda_aws_token.sh -r ~/.nda_credential)"
+            }
+            $AWS s3 cp --no-progress $LOC $SM/downloads/ && { rc=$?; break; } || rc=$?
+        else
+            ls -lh $LOC && ln -sf $(readlink -f $LOC) $SM/downloads/ || rc=$?
+            break
+        fi
+        n=$((n+1))
+        printf "Download try $n failed.\n\n"
+    done
+    [[ $rc -eq 0 ]] || false
+    touch $DONE
+fi
 
 printf -- "[$(date)] Finish downlaod: $FNAME\n---\n"
